APIMF-3529: move generation of source information node from AMFCompiler to transformation step
