+++ shared/src/main/scala/amf/core/internal/parser/AMFCompiler.scala
@@ -10,7 +10,6 @@ import amf.core.client.scala.parse.TaggedReferences._
 import amf.core.client.scala.parse.document.{UnresolvedReference => _, _}
 import amf.core.internal.remote.Mimes._
 import amf.core.internal.adoption.IdAdopter
-import amf.core.internal.parser.domain.SourceInformationGenerator
 import amf.core.internal.remote._
 import amf.core.internal.utils.AmfStrings
 import amf.core.internal.validation.CoreValidations._
@@ -119,12 +118,8 @@ class AMFCompiler(compilerContext: CompilerContext, val referenceKind: Reference
       if (document.location == compilerContext.fileContext.root) baseUnit.withRoot(true)
       baseUnit.withRaw(document.raw).tagReferences(documentWithReferences)
       if (isRoot && domainPlugin.withIdAdoption) new IdAdopter(baseUnit, document.location).adoptFromRoot()
-      if (isRoot) SourceInformationGenerator.generateSourceInformation(baseUnit, documentWithReferences.location)
+      parsedModelEvent(baseUnit)
       baseUnit
-    } map { unit =>
-      // we setup the run for the parsed unit
-      parsedModelEvent(unit)
-      unit
     }
   }
